+++ stories/helpers/createApiStory.js
@@ -3,7 +3,7 @@ import React from 'react'
 const addRefToInfo = (info = '') =>
   info +
   `\n
-The \`ref\` will be forwared to the **html** root element, as well as any other props.
+The \`ref\` will be forwarded to the **html** root element, as well as any other props.
 `
 
 export default (story, Component, info) => {